consolidate data intake CLI into single script
