+++ src/epu_data_intake/README.md
@@ -1,5 +1,3 @@
-from src.smartem_decisions.epu_data_intake.parser import EpuSession
-
 # EPU directory parsing and data intake
 
 An EPU data directory is generated by closed-source EPU software and represents an acquisition session
@@ -17,28 +15,29 @@ using a cryo-electron microscope. Data intake scripts can be used to:
 
 ```bash
 # parse complete EPU directory
-python src/epu_data_intake/fs_parser.py parse-epu-dir \
+python -m epu_data_intake parse dir \
   ../metadata_Supervisor_20250108_101446_62_cm40593-1_EPU
 
-python src/epu_data_intake/fs_parser.py parse-epu-dir \
+python -m epu_data_intake parse dir \
   ../metadata_Supervisor_20250114_220855_23_epuBSAd20_GrOxDDM
 
-python src/epu_data_intake/fs_parser.py parse-epu-dir \
+python -m epu_data_intake parse dir \
   ../metadata_Supervisor_20241220_140307_72_et2_gangshun
 
 # parse things
-python src/epu_data_intake/fs_parser.py parse-epu-session \
+python -m epu_data_intake parse session \
   tests/testdata/epu-dir-example/EpuSession.dm
 
-python src/epu_data_intake/fs_parser.py parse-atlas \
+python -m epu_data_intake parse atlas \
   tests/testdata/Atlas.dm
 
-python src/epu_data_intake/fs_parser.py parse-gridsquare \
+python -m epu_data_intake parse gridsquare \
   tests/testdata/epu-dir-example/Images-Disc1/GridSquare_8999138/GridSquare_20250108_151151.xml
 
-python src/epu_data_intake/fs_parser.py parse-gridsquare-metadata \
+python -m epu_data_intake parse gridsquare-metadata \
   tests/testdata/epu-dir-example/Metadata/GridSquare_8999138.dm
-python src/epu_data_intake/fs_parser.py parse-gridsquare-metadata \
+
+python -m epu_data_intake parse gridsquare-metadata \
   ./tests/testdata/bi37708-28/Supervisor_20250129_134723_36_bi37708-28_grid7_EPU/Metadata/GridSquare_29273435.dm
 # ../metadata_Supervisor_20241220_140307_72_et2_gangshun/Metadata/GridSquare_20688814.dm
 # ../metadata_Supervisor_20241220_140307_72_et2_gangshun/Metadata/GridSquare_20688837.dm
@@ -50,43 +49,43 @@ python src/epu_data_intake/fs_parser.py parse-gridsquare-metadata \
 # ../metadata_Supervisor_20250114_220855_23_epuBSAd20_GrOxDDM/Metadata/GridSquare_13515293.dm
 # ../metadata_Supervisor_20250114_220855_23_epuBSAd20_GrOxDDM/Metadata/GridSquare_13515237.dm
 
-python src/epu_data_intake/fs_parser.py parse-foilhole \
+python -m epu_data_intake parse foilhole \
   tests/testdata/epu-dir-example/Images-Disc1/GridSquare_8999138/FoilHoles/FoilHole_9015889_20250108_154725.xml
 
-python src/epu_data_intake/fs_parser.py parse-micrograph \
+python -m epu_data_intake parse micrograph \
   tests/testdata/epu-dir-example/Images-Disc1/GridSquare_8999138/Data/FoilHole_9015883_Data_9017354_6_20250108_154918.xml
   
 # Validate epu project dirs:
 
 # expect failure:
-python src/epu_data_intake/fs_parser.py validate-epu-dir \
+python -m epu_data_intake validate \
   tests/testdata/bi37708-28/Supervisor_20250129_114842_73_bi37708-28_grid7_EPU
 
 # expect success:
-python src/epu_data_intake/fs_parser.py validate-epu-dir \
+python -m epu_data_intake validate \
   tests/testdata/bi37708-28/Supervisor_20250129_134723_36_bi37708-28_grid7_EPU
 
 # expect failure:
-python src/epu_data_intake/fs_parser.py validate-epu-dir \
+python -m epu_data_intake validate \
   tests/testdata/bi37708-28/Supervisor_20250130_105058_11
 
 # expect success:
-python src/epu_data_intake/fs_parser.py validate-epu-dir \
+python -m epu_data_intake validate \
   tests/testdata/bi37708-28/Supervisor_20250130_133418_68apoferritin
 
-python src/epu_data_intake/fs_parser.py validate-epu-dir \
+python -m epu_data_intake validate \
   tests/testdata/bi37708-28/Supervisor_20250130_143856_44Practice
 
 # expect success:
-python src/epu_data_intake/fs_parser.py validate-epu-dir \
+python -m epu_data_intake validate \
   tests/testdata/bi37708-28/Supervisor_20250130_145409_68
 
 # expect success:
-python src/epu_data_intake/fs_parser.py validate-epu-dir \
+python -m epu_data_intake validate \
   tests/testdata/bi37708-28/Supervisor_20250130_145409_68practice2
 
 # expect failure:
-python src/epu_data_intake/fs_parser.py validate-epu-dir \
+python -m epu_data_intake validate \
   tests/testdata/bi37708-28/Supervisor_20250130_150924_1grid3
 
 ```
@@ -194,7 +193,7 @@ We need to watch the EPU output directory for changes, to that end:
 ```bash
 # Launch the watcher:
 rm -f -- output.log && rm -rf ../fs-watcher-test-dir/* && \
-  python src/epu_data_intake/fs_watcher.py ../fs-watcher-test-dir/ \
+  python -m epu_data_intake watch ../fs-watcher-test-dir/ \
   --log-file output.log
 
 # In another terminal launch a simulator that randomly writes to target dir:
